Remove Frame enum
Refactor things up a bit to use a `TypedMessage` struct instead of the
enum, which contains the `MessageType` and the message as a `Box<dyn
Message>`. We'll see is that makes things simpler or not...

Also added an `AsyncTypedMessage` that adds a `oneshot::Receiver<()` so
the caller can be notified when a message has been successfully sent to
the peer.