Start using the model and send Index message
Right now the model uses interior mutability and is shared accross
connections using an `Arc`, which is not quite ideal. It might need to
be spawned into its own task and interacted with via message passing at
some point.

We now send an (empty for now) `Index` message to the peer. Next step is
to request various files.